调整基础类字段定义

+++ shrimp-cloud-mybatis/src/main/java/com/wkclz/mybatis/base/BaseMapper.java
@@ -16,7 +16,7 @@ import java.util.List;
 public interface BaseMapper<Entity> {
 
     @Desc("统计")
-    Integer count(Entity entity);
+    Long count(Entity entity);
 
     @Desc("用ID查找")
     Entity getById(@Param("id") Long id);
